extract provided type
