+++ src/find.ts
@@ -1,6 +1,6 @@
 import { toSingle } from "./internal/toSingle";
+import type { LazyEvaluator } from "./internal/types/LazyEvaluator";
 import { SKIP_ITEM } from "./internal/utilityEvaluators";
-import type { LazyEvaluator } from "./pipe";
 import { purry } from "./purry";
 
 /**
