+++ src/first.ts
@@ -1,6 +1,6 @@
 import { toSingle } from "./internal/toSingle";
 import type { IterableContainer } from "./internal/types/IterableContainer";
-import type { LazyEvaluator } from "./pipe";
+import type { LazyEvaluator } from "./internal/types/LazyEvaluator";
 import { purry } from "./purry";
 
 type First<T extends IterableContainer> = T extends []
