+++ src/internal/types/LazyEvaluator.ts
@@ -1,0 +1,7 @@
+import type { LazyResult } from "./LazyResult";
+
+export type LazyEvaluator<T = unknown, R = T> = (
+  item: T,
+  index: number,
+  data: ReadonlyArray<T>,
+) => LazyResult<R>;
