+++ src/unique.ts
@@ -1,8 +1,8 @@
 import { purryFromLazy } from "./internal/purryFromLazy";
 import type { Deduped } from "./internal/types/Deduped";
 import type { IterableContainer } from "./internal/types/IterableContainer";
+import type { LazyEvaluator } from "./internal/types/LazyEvaluator";
 import { SKIP_ITEM } from "./internal/utilityEvaluators";
-import type { LazyEvaluator } from "./pipe";
 
 /**
  * Returns a new array containing only one copy of each element in the original
