+++ src/zip.ts
@@ -1,5 +1,5 @@
 import type { IterableContainer } from "./internal/types/IterableContainer";
-import type { LazyEvaluator } from "./pipe";
+import type { LazyEvaluator } from "./internal/types/LazyEvaluator";
 import { purry } from "./purry";
 
 type Zipped<Left extends IterableContainer, Right extends IterableContainer> =
